requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.1
email-validator==2.2.0
google-generativeai==0.8.3
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
beautifulsoup4==4.12.3
lxml==5.1.0

